+++ tests/test_minimal/test_tools/test_backend_and_dataset_configuration/test_helpers/test_configure_backend_overrides.py
@@ -5,14 +5,14 @@ from typing import Literal
 import numcodecs
 import numpy as np
 import pytest
+from hdmf.common import DynamicTable, VectorData
 from hdmf.data_utils import DataChunkIterator
-from hdmf_zarr import NWBZarrIO
-from pynwb import NWBHDF5IO
 from pynwb.testing.mock.base import mock_TimeSeries
 from pynwb.testing.mock.file import mock_NWBFile
 
 from neuroconv.tools.hdmf import SliceableDataChunkIterator
 from neuroconv.tools.nwb_helpers import (
+    BACKEND_NWB_IO,
     configure_backend,
     get_default_backend_configuration,
 )
@@ -31,8 +31,6 @@ from neuroconv.tools.nwb_helpers import (
 def test_simple_time_series_override(
     tmpdir: Path, case_name: str, iterator: callable, iterator_options: dict, backend: Literal["hdf5", "zarr"]
 ):
-    BACKEND_NWB_IO = dict(hdf5=NWBHDF5IO, zarr=NWBZarrIO)
-
     array = np.zeros(shape=(30_000 * 5, 384), dtype="int16")
     data = iterator(array, **iterator_options)
 
@@ -59,13 +57,70 @@ def test_simple_time_series_override(
     if case_name != "unwrapped":  # TODO: eventually, even this case will be buffered automatically
         assert nwbfile.acquisition["TestTimeSeries"].data
 
-    nwbfile_path = str(tmpdir / f"test_configure_{backend}_defaults_{case_name}_data.nwb.h5")
+    nwbfile_path = str(tmpdir / f"test_configure_defaults_{case_name}_data.nwb.{backend}")
     with BACKEND_NWB_IO[backend](path=nwbfile_path, mode="w") as io:
         io.write(nwbfile)
 
     with BACKEND_NWB_IO[backend](path=nwbfile_path, mode="r") as io:
         written_nwbfile = io.read()
         written_data = written_nwbfile.acquisition["TestTimeSeries"].data
+
+        assert written_data.chunks == smaller_chunk_shape
+
+        if backend == "hdf5":
+            assert written_data.compression == "gzip"
+            assert written_data.compression_opts == higher_gzip_level
+        elif backend == "zarr":
+            assert written_data.compressor == numcodecs.GZip(level=5)
+
+
+@pytest.mark.parametrize(
+    "case_name,iterator,iterator_options",
+    [
+        ("unwrapped", lambda x: x, dict()),
+        ("generic", SliceableDataChunkIterator, dict()),
+        ("classic", DataChunkIterator, dict(iter_axis=1, buffer_size=30_000 * 5)),
+        # Need to hardcode buffer size in classic case or else it takes forever...
+    ],
+)
+@pytest.mark.parametrize("backend", ["hdf5", "zarr"])
+def test_simple_dynamic_table_override(
+    tmpdir: Path, case_name: str, iterator: callable, iterator_options: dict, backend: Literal["hdf5", "zarr"]
+):
+    data = np.zeros(shape=(30_000 * 5, 384), dtype="int16")
+
+    nwbfile = mock_NWBFile()
+    dynamic_table = DynamicTable(
+        name="TestDynamicTable", description="", columns=[VectorData(name="TestColumn", description="", data=data)]
+    )
+    nwbfile.add_acquisition(dynamic_table)
+
+    backend_configuration = get_default_backend_configuration(nwbfile=nwbfile, backend=backend)
+    dataset_configuration = backend_configuration.dataset_configurations["acquisition/TestDynamicTable/TestColumn/data"]
+
+    smaller_chunk_shape = (30_000, 64)
+    smaller_buffer_shape = (60_000, 192)
+    dataset_configuration.chunk_shape = smaller_chunk_shape
+    dataset_configuration.buffer_shape = smaller_buffer_shape
+
+    higher_gzip_level = 5
+    if backend == "hdf5":
+        dataset_configuration.compression_options = dict(level=higher_gzip_level)
+    elif backend == "zarr":
+        dataset_configuration.compression_options = dict(level=higher_gzip_level)
+
+    configure_backend(nwbfile=nwbfile, backend_configuration=backend_configuration)
+
+    if case_name != "unwrapped":  # TODO: eventually, even this case will be buffered automatically
+        assert nwbfile.acquisition["TestDynamicTable"]["TestColumn"].data
+
+    nwbfile_path = str(tmpdir / f"test_configure_defaults_{case_name}_dynamic_table.nwb.{backend}")
+    with BACKEND_NWB_IO[backend](path=nwbfile_path, mode="w") as io:
+        io.write(nwbfile)
+
+    with BACKEND_NWB_IO[backend](path=nwbfile_path, mode="r") as io:
+        written_nwbfile = io.read()
+        written_data = written_nwbfile.acquisition["TestDynamicTable"]["TestColumn"].data
 
         assert written_data.chunks == smaller_chunk_shape